Changed user id to be unique unsigned integer in tutors table.

+++ server/database/migrations/1617425021722_subjects.ts
@@ -15,6 +15,6 @@ export default class Subjects extends BaseSchema {
     }
   }
   public async down() {
-    this.schema.dropTable(this.tableName)
+    if (await this.schema.hasTable(this.tableName)) this.schema.dropTable(this.tableName)
   }
 }
